+++ gherkin/go/matcher.go
@@ -92,22 +92,36 @@ func (m *matcher) MatchComment(line *Line) (ok bool, token *Token, err error) {
 }
 
 func (m *matcher) MatchTagLine(line *Line) (ok bool, token *Token, err error) {
-	if line.StartsWith(TAG_PREFIX) {
-		var tags []*LineSpan
-		var column = line.Indent()
-		splits := strings.Split(line.TrimmedLineText, TAG_PREFIX)
-		for i := range splits {
-			txt := strings.Trim(splits[i], " ")
-			if txt != "" {
-				tags = append(tags, &LineSpan{column, TAG_PREFIX + txt})
-			}
-			column = column + utf8.RuneCountInString(splits[i]) + 1
-		}
-
-		token, ok = m.newTokenAtLocation(line.LineNumber, line.Indent()), true
-		token.Type = TokenTypeTagLine
-		token.Items = tags
+	if !line.StartsWith(TAG_PREFIX) {
+		return
 	}
+	commentDelimiter := regexp.MustCompile(`\s+` + COMMENT_PREFIX)
+	uncommentedLine := commentDelimiter.Split(line.TrimmedLineText, 2)[0]
+	var tags []*LineSpan
+	var column = line.Indent() + 1
+
+	splits := strings.Split(uncommentedLine, TAG_PREFIX)
+	for i := range splits {
+		txt := strings.TrimRightFunc(splits[i], func(r rune) bool {
+			return unicode.IsSpace(r)
+		})
+		if len(txt) == 0 {
+			continue
+		}
+		if !regexp.MustCompile(`^\S+$`).MatchString(txt) {
+			location := &Location{line.LineNumber, column}
+			msg := "A tag may not contain whitespace"
+			err = &parseError{msg, location}
+			break
+		}
+		tags = append(tags, &LineSpan{column, TAG_PREFIX + txt})
+		column = column + utf8.RuneCountInString(splits[i]) + 1
+	}
+
+	token, ok = m.newTokenAtLocation(line.LineNumber, line.Indent()), true
+	token.Type = TokenTypeTagLine
+	token.Items = tags
+
 	return
 }
 
@@ -190,6 +204,10 @@ func (m *matcher) MatchDocStringSeparator(line *Line) (ok bool, token *Token, er
 	return
 }
 
+func isSpaceAndNotNewLine(r rune) bool {
+	return unicode.IsSpace(r) && r != '\n'
+}
+
 func (m *matcher) MatchTableRow(line *Line) (ok bool, token *Token, err error) {
 	var firstChar, firstPos = utf8.DecodeRuneInString(line.TrimmedLineText)
 	if firstChar == TABLE_CELL_SEPARATOR {
@@ -204,9 +222,9 @@ func (m *matcher) MatchTableRow(line *Line) (ok bool, token *Token, err error) {
 				// append current cell
 				txt := string(cell)
 
-				txtTrimmedLeadingSpace := strings.TrimLeftFunc(txt, unicode.IsSpace)
+				txtTrimmedLeadingSpace := strings.TrimLeftFunc(txt, isSpaceAndNotNewLine)
 				ind := utf8.RuneCountInString(txt) - utf8.RuneCountInString(txtTrimmedLeadingSpace)
-				txtTrimmed := strings.TrimRightFunc(txtTrimmedLeadingSpace, unicode.IsSpace)
+				txtTrimmed := strings.TrimRightFunc(txtTrimmedLeadingSpace, isSpaceAndNotNewLine)
 				cells = append(cells, &LineSpan{startCol + ind, txtTrimmed})
 				// start building next
 				cell = make([]rune, 0)
@@ -272,9 +290,11 @@ func (m *matcher) MatchOther(line *Line) (ok bool, token *Token, err error) {
 }
 
 func (m *matcher) unescapeDocString(text string) string {
-	if m.activeDocStringSeparator != "" {
-		return strings.Replace(text, "\\\"\\\"\\\"", "\"\"\"", -1)
-	} else {
-		return text
+	if m.activeDocStringSeparator == DOCSTRING_SEPARATOR {
+		return strings.Replace(text, "\\\"\\\"\\\"", DOCSTRING_SEPARATOR, -1)
 	}
+	if m.activeDocStringSeparator == DOCSTRING_ALTERNATIVE_SEPARATOR {
+		return strings.Replace(text, "\\`\\`\\`", DOCSTRING_ALTERNATIVE_SEPARATOR, -1)
+	}
+	return text
 }