+++ fake-cucumber/javascript/src/TestStep.ts
@@ -29,7 +29,7 @@ export default abstract class TestStep implements ITestStep {
     world: IWorld,
     notifier: MessageNotifier,
     testCaseStartedId: string
-  ): Promise<messages.ITestResult> {
+  ): Promise<messages.ITestStepResult> {
     this.emitTestStepStarted(notifier, testCaseStartedId)
 
     const start = this.clock.now()
@@ -39,9 +39,9 @@ export default abstract class TestStep implements ITestStep {
 
       return this.emitTestStepFinished(
         testCaseStartedId,
-        new messages.TestResult({
+        new messages.TestStepResult({
           duration: duration,
-          status: messages.TestResult.Status.UNDEFINED,
+          status: messages.TestStepResult.Status.UNDEFINED,
         }),
         notifier
       )
@@ -52,9 +52,9 @@ export default abstract class TestStep implements ITestStep {
 
       return this.emitTestStepFinished(
         testCaseStartedId,
-        new messages.TestResult({
+        new messages.TestStepResult({
           duration: duration,
-          status: messages.TestResult.Status.AMBIGUOUS,
+          status: messages.TestStepResult.Status.AMBIGUOUS,
         }),
         notifier
       )
@@ -71,12 +71,12 @@ export default abstract class TestStep implements ITestStep {
       const duration = millisecondsToDuration(finish - start)
       return this.emitTestStepFinished(
         testCaseStartedId,
-        new messages.TestResult({
+        new messages.TestStepResult({
           duration,
           status:
             result === 'pending'
-              ? messages.TestResult.Status.PENDING
-              : messages.TestResult.Status.PASSED,
+              ? messages.TestStepResult.Status.PENDING
+              : messages.TestStepResult.Status.PASSED,
         }),
         notifier
       )
@@ -87,9 +87,9 @@ export default abstract class TestStep implements ITestStep {
       const duration = millisecondsToDuration(finish - start)
       return this.emitTestStepFinished(
         testCaseStartedId,
-        new messages.TestResult({
+        new messages.TestStepResult({
           duration,
-          status: messages.TestResult.Status.FAILED,
+          status: messages.TestStepResult.Status.FAILED,
           message,
         }),
         notifier
@@ -100,13 +100,13 @@ export default abstract class TestStep implements ITestStep {
   public skip(
     notifier: MessageNotifier,
     testCaseStartedId: string
-  ): messages.ITestResult {
+  ): messages.ITestStepResult {
     this.emitTestStepStarted(notifier, testCaseStartedId)
     return this.emitTestStepFinished(
       testCaseStartedId,
-      new messages.TestResult({
+      new messages.TestStepResult({
         duration: millisecondsToDuration(0),
-        status: messages.TestResult.Status.SKIPPED,
+        status: messages.TestStepResult.Status.SKIPPED,
       }),
       notifier
     )
@@ -129,19 +129,19 @@ export default abstract class TestStep implements ITestStep {
 
   protected emitTestStepFinished(
     testCaseStartedId: string,
-    testResult: messages.ITestResult,
+    testStepResult: messages.ITestStepResult,
     notifier: MessageNotifier
-  ): messages.ITestResult {
+  ): messages.ITestStepResult {
     notifier(
       new messages.Envelope({
         testStepFinished: new messages.TestStepFinished({
           testCaseStartedId,
           testStepId: this.id,
-          testResult,
+          testStepResult,
           timestamp: millisecondsSinceEpochToTimestamp(this.clock.now()),
         }),
       })
     )
-    return testResult
+    return testStepResult
   }
 }