+++ json-formatter/CHANGELOG.md
@@ -21,6 +21,8 @@ This document is formatted according to the principles of [Keep A CHANGELOG](htt
 
 * Process rules
 
+* Output background name
+
 ## [2.2.0] - 2020-01-10
 
 ### Changed