+++ json-formatter/CHANGELOG.md
@@ -20,6 +20,7 @@ This document is formatted according to the principles of [Keep A CHANGELOG](htt
 ### Fixed
 
 * Process rules
+* Output background name
 
 ## [2.2.0] - 2020-01-10